requirements.txt
streamlit>=1.30.0
openpyxl>=3.1.0
pypdf>=3.0.0
Pillow>=10.0.0
fpdf2>=2.7.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
